refactor(icons): move icons to root src directory
