+++ packages/icons/src/blockquote.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Blockquote: IconDefinition = {
   name: 'blockquote',
