+++ packages/icons/src/bold.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Bold: IconDefinition = {
   name: 'bold',
