+++ packages/icons/src/erase.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Erase: IconDefinition = {
   name: 'erase',
