+++ packages/icons/src/facebook.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Facebook: IconDefinition = {
   name: 'facebook',
