+++ packages/icons/src/heading1.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Heading1: IconDefinition = {
   name: 'heading1',
