+++ packages/icons/src/highlight.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Highlight: IconDefinition = {
   name: 'highlight',
