+++ packages/icons/src/image.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Image: IconDefinition = {
   name: 'image',
