+++ packages/icons/src/instagram.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Instagram: IconDefinition = {
   name: 'instagram',
