+++ packages/icons/src/italic.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Italic: IconDefinition = {
   name: 'italic',
