+++ packages/icons/src/link.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Link: IconDefinition = {
   name: 'link',
