+++ packages/icons/src/underline.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Underline: IconDefinition = {
   name: 'underline',
