+++ packages/icons/src/unlink.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Unlink: IconDefinition = {
   name: 'unlink',
