+++ packages/icons/src/video.ts
@@ -1,4 +1,4 @@
-import { IconDefinition } from '../types';
+import { IconDefinition } from './types';
 
 export const Video: IconDefinition = {
   name: 'video',
